cephadm: create deployment_utils module
Create a deployment_utils module for deployment related functions that
don't have a better home.
